🚧 primeira versão aceitável
