+++ src/pages/Home/Banners/Banner/styles.js
@@ -1,19 +1,13 @@
 import { Card } from 'react-native-paper';
 import styled from 'styled-components/native';
 
-export const ConteudoImagem = styled.View`
-  width: 100%;
-  height: 130px;
-`;
-
-export const Imagem = styled.Image`
-  border-radius: 10px;
-  height: 130px;
-  width: 100%;
-`;
-
-export const Cartao = styled(Card)`
+export const Container = styled(Card)`
   height: 130px;
   border-radius: 10px;
   margin: 20px 16px;
+`;
+
+export const ConteudoImagem = styled.View`
+  width: 100%;
+  height: 130px;
 `;
