--- src/pages/Home/Banners/NewBanner/styles.js
@@ -1,13 +1,0 @@
-import { Card } from 'react-native-paper';
-import styled from 'styled-components/native';
-
-export const Container = styled(Card)`
-  height: 130px;
-  border-radius: 10px;
-  margin: 20px 16px;
-`;
-
-export const ConteudoImagem = styled.View`
-  width: 100%;
-  height: 130px;
-`;
